Added full NES assembler.

+++ src/main/kotlin/io/github/gaeqs/nes4jams/utils/IntRanges.kt
@@ -1,4 +1,5 @@
 package io.github.gaeqs.nes4jams.utils
 
-val BYTE_RANGE = Byte.MIN_VALUE..Byte.MAX_VALUE
-val SHORT_RANGE = Short.MIN_VALUE..Short.MAX_VALUE
+// The range of possible number the user can enter is from the minimum signed value to the maximum unsigned value!
+val BYTE_RANGE = Byte.MIN_VALUE..UByte.MAX_VALUE.toInt()
+val SHORT_RANGE = Short.MIN_VALUE..UShort.MAX_VALUE.toInt()
